DE6425 -- hosLog Records have a 64 bit (long) id, and code was set up for a 32 bit (int) id -- Please note: the hos.sql script will also need to be run on deploy (it has been run on the slb test db, dev, and qa)

+++ dao/src/main/java/com/inthinc/pro/dao/HOSDAO.java
@@ -14,7 +14,7 @@ import com.inthinc.pro.model.hos.HOSRecord;
 import com.inthinc.pro.model.hos.HOSVehicleDayData;
 import com.inthinc.pro.model.hos.HOSVehicleMileage;
 
-public interface HOSDAO extends GenericDAO<HOSRecord, Integer> {
+public interface HOSDAO extends GenericDAO<HOSRecord, Long> {
     
     List<HOSRecord> getHOSRecords(Integer driverID, Interval interval, Boolean driverStatusOnly);
     List<HOSRecord> getHOSRecordsFilteredByInterval(Integer driverID, Interval interval, Boolean driverStatusOnly);
